courier id change from int to long

+++ src/main/java/com/ataberkcanitez/migroscouriertracking/domain/common/observer/CourierLocationSubscriber.java
@@ -3,5 +3,5 @@ package com.ataberkcanitez.migroscouriertracking.domain.common.observer;
 import com.ataberkcanitez.migroscouriertracking.domain.location.model.Location;
 
 public interface CourierLocationSubscriber {
-    void receiveLocationUpdate(int courierId, Location location);
+    void receiveLocationUpdate(long courierId, Location location);
 }
